requirements.txt
boto3==1.28.81
botocore==1.31.81
ansible==8.6.0
python-dotenv==1.0.0
